requirements.txt
# Discord bot dependencies
discord.py>=2.3.0
aiohttp>=3.8.0
requests>=2.28.0

# GUI dependencies
customtkinter>=5.2.0
pillow>=10.0.0

# Build dependencies
pyinstaller>=6.0.0
